migrate test_nftables to pytest
